Refactored DashboardController and added relationships to User Model
The DashboardController was updated to use new models CompletedTask and Leaderboard instead of the DB facade, to achieve an object-oriented approach enhancing maintainability. User model has been extended with a 'hasOne' relationship to the Leaderboard and an 'onCreate' event to initialize a leaderboard instance with 0 score for a new user. Points are no longer stored in the CompletedTask model, as they are now tracked in the Leaderboard model.