Add dompdf package for PDF generation
The barryvdh/laravel-dompdf package was added to the composer.json file. This package helps in generating PDF files from HTML. It was necessary for creating a feature that allows users to download and print the leaderboard as a PDF file. Other dependencies of the package were also added in the composer.lock file.