Add leaderboard export to PDF and CSV functionality
A new functionality was added to the LeaderboardController, 'exportToPDF' and 'exportToCSV', which allows users to export the leaderboard data as PDF or CSV. The respective routes were also defined in the routes/web.php file. This enhances the usability of the leaderboard feature by providing flexible data export options. Commit includes addition of PDF generation library Barryvdh\DomPDF.