Add 'solve_count' feature to tasks
'Solve_count' field is added to 'tasks' table to keep track of the number of times a task was solved. This change is made to seize statistics about each task. Also, dashboard UI is updated to display 'solve_count' for each task. Task model and TasksController are updated accordingly to manage 'solve_count' field.