Update UI and add solve count to tasks
This commit refers to UI changes made to enhance legibility by adjusting task category size and title truncation. In addition, a 'solve_count' feature is introduced to document the number of times a task has been completed. This data will serve to provide a useful statistic for each task. Necessary adjustments have been made to the Task model and TasksController to accommodate this feature.