Increase margin-bottom in task view
The margin-bottom of the container in the task view was increased for better visual separation from consecutive elements. The change was made in the task.blade.php file in the 'div' class by increasing the margin-bottom from 'mb-40' to 'mb-80'. This extra spacing makes the interface more user-friendly and easier on the eyes.